Move all group related servlets into rest/admin/groups.py
